Now handle the possibility of a single empty bin during gibbs sampling. Split largest bin and assign half the points to empty bin.
It is possible that the probability of assigning a data point to a specific bin may be zero, or so small that sampling from the probability distribution over the bin indices may cause a bin index to never be selected for. This appears to typically result in at most 1 empty bin (although technically could be more...). This would cause flow on issues when calculating alpha values (as there would be no counts in the bin in question). By assigning half the data points to the empty bin, we should end up with two components with roughly the same centers. An extension to this fix would account for the possibility of multiple empty bins.